# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ErrorInfo'
      message:
        example: Success
        type: string
      status:
        example: true
        type: boolean
    type: object
  models.ErrorInfo:
    properties:
      code:
        example: ERROR_CODE
        type: string
      details:
        example: Detailed error message
        type: string
    type: object
  models.FeaturedMoviesResponse:
    properties:
      movies:
        items:
          $ref: '#/definitions/models.Movie'
        type: array
    type: object
  models.Movie:
    properties:
      poster:
        example: https://image.tmdb.org/t/p/w185/poster.jpg
        type: string
      title:
        example: Example Movie (2024)
        type: string
      type:
        example: movie
        type: string
      url:
        example: https://tv12.idlixku.com/movie/example-2024/
        type: string
      year:
        example: "2024"
        type: string
    type: object
  models.SubtitleInfo:
    properties:
      available:
        example: true
        type: boolean
      tracks:
        items:
          $ref: '#/definitions/models.SubtitleTrack'
        type: array
    type: object
  models.SubtitleSearchResponse:
    properties:
      filtered:
        example: true
        type: boolean
      subtitles:
        items:
          $ref: '#/definitions/models.SubtitleTrackInfo'
        type: array
      total:
        example: 2
        type: integer
      video_id:
        example: "163426"
        type: string
      video_name:
        example: Crime 101 (2026)
        type: string
    type: object
  models.SubtitleTrack:
    properties:
      format:
        example: srt
        type: string
      language:
        example: Bahasa
        type: string
      url:
        example: https://g5.wiseacademia.asia/r/xyz.jpg
        type: string
    type: object
  models.SubtitleTrackInfo:
    properties:
      download_url:
        example: /api/v1/subtitle/download?url=...
        type: string
      format:
        example: srt
        type: string
      language:
        example: Bahasa
        type: string
      url:
        example: https://g5.wiseacademia.asia/r/xyz.jpg
        type: string
    type: object
  models.VariantPlaylist:
    properties:
      bandwidth:
        example: 1510000
        type: integer
      id:
        example: "0"
        type: string
      resolution:
        example: 1920x1080
        type: string
      uri:
        example: https://jeniusplay.com/hls/hash123/1080p.m3u8
        type: string
    type: object
  models.VideoInfo:
    properties:
      embed_url:
        example: https://jeniusplay.com/video/hash123
        type: string
      is_variant_playlist:
        example: true
        type: boolean
      m3u8_url:
        example: https://jeniusplay.com/cdn/hls/hash123/master.m3u8
        type: string
      poster:
        example: https://image.tmdb.org/t/p/w185/poster.jpg
        type: string
      subtitle:
        $ref: '#/definitions/models.SubtitleInfo'
      variants:
        items:
          $ref: '#/definitions/models.VariantPlaylist'
        type: array
      video_id:
        example: "163426"
        type: string
      video_name:
        example: Crime 101 (2026)
        type: string
    type: object
  models.VideoInfoRequest:
    properties:
      url:
        example: https://tv12.idlixku.com/movie/crime-101-2026/
        type: string
    required:
    - url
    type: object
host: localhost:8080
info:
  contact:
    email: nabilulilalbab@github.com
    name: API Support
    url: https://github.com/nabilulilalbab/streamv2/issues
  description: RESTful API for IDLIX video streaming platform. Provides endpoints
    for scraping movie data, M3U8 streams, and video variants.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/nabilulilalbab/streamv2
  title: IDLIX API
  version: 1.0.0
paths:
  /featured:
    get:
      consumes:
      - application/json
      description: Retrieve list of featured movies from IDLIX homepage (excluding
        TV series)
      produces:
      - application/json
      responses:
        "200":
          description: Featured movies retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.FeaturedMoviesResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get featured movies
      tags:
      - movies
  /health:
    get:
      consumes:
      - application/json
      description: Check if API is running and get version information
      produces:
      - application/json
      responses:
        "200":
          description: API is healthy
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - system
  /proxy:
    get:
      description: Proxy streaming requests to bypass CORS restrictions
      parameters:
      - description: Target URL to proxy
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/x-mpegURL
      - video/MP2T
      responses:
        "200":
          description: Stream content
          schema:
            type: string
        "400":
          description: Invalid URL
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Proxy error
          schema:
            additionalProperties: true
            type: object
      summary: Proxy M3U8/TS requests
      tags:
      - proxy
  /subtitle/download:
    get:
      description: Download subtitle file with CORS support and optional format conversion
      parameters:
      - description: Subtitle URL
        in: query
        name: url
        required: true
        type: string
      - description: Target format (srt or vtt)
        in: query
        name: format
        type: string
      - description: Custom filename
        in: query
        name: filename
        type: string
      produces:
      - text/plain
      - text/vtt
      responses:
        "200":
          description: Subtitle file content
          schema:
            type: string
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Download or conversion error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Download subtitle file
      tags:
      - subtitle
  /subtitle/search:
    get:
      consumes:
      - application/json
      description: Get subtitle tracks for a video with optional language filter
      parameters:
      - description: Video URL
        in: query
        name: url
        required: true
        type: string
      - description: Language filter (case-insensitive, partial match)
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subtitles found
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SubtitleSearchResponse'
              type: object
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Failed to get subtitles
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Search subtitle tracks
      tags:
      - subtitle
  /video/info:
    post:
      consumes:
      - application/json
      description: Get complete video information including M3U8 streams, quality
        variants, and subtitles
      parameters:
      - description: Video URL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VideoInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Video info retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.VideoInfo'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get video information
      tags:
      - videos
schemes:
- http
- https
swagger: "2.0"
tags:
- description: System health and status endpoints
  name: system
- description: Movie browsing and discovery endpoints
  name: movies
- description: Video information and streaming endpoints
  name: videos
- description: Subtitle download and search endpoints
  name: subtitle